requirements.txt
flask==3.1.0
flask-cors==5.0.1
PyPDF2==3.0.1
python-docx==1.1.2
anthropic==0.43.0
python-dotenv==1.1.0
gunicorn==23.0.0
